+++ Kernel/Output/PDF/Ticket.pm
@@ -1,5 +1,6 @@
 # --
 # Copyright (C) 2001-2021 OTRS AG, https://otrs.com/
+# Copyright (C) 2021 Znuny GmbH, https://znuny.org/
 # --
 # This software comes with ABSOLUTELY NO WARRANTY. For details, see
 # the enclosed file COPYING for license information (GPL). If you
@@ -20,14 +21,13 @@ our @ObjectDependencies = (
     'Kernel::System::DynamicField::Backend',
     'Kernel::System::LinkObject',
     'Kernel::System::Log',
-    'Kernel::System::JSON',
     'Kernel::System::PDF',
-    'Kernel::System::User',
     'Kernel::System::Ticket',
     'Kernel::System::Ticket::Article',
+    'Kernel::System::User',
 );
 
-use Kernel::System::VariableCheck qw(IsHashRefWithData);
+use Kernel::System::VariableCheck qw(:all);
 
 sub new {
     my ( $Type, %Param ) = @_;
@@ -860,7 +860,10 @@ sub _PDFOutputTicketDynamicFields {
 sub _PDFOutputCustomerInfos {
     my ( $Self, %Param ) = @_;
 
-    # Check needed stuff.
+    my $LayoutObject              = $Kernel::OM->Get('Kernel::Output::HTML::Layout');
+    my $DynamicFieldObject        = $Kernel::OM->Get('Kernel::System::DynamicField');
+    my $DynamicFieldBackendObject = $Kernel::OM->Get('Kernel::System::DynamicField::Backend');
+
     for my $Needed (qw(PageData CustomerData)) {
         if ( !defined( $Param{$Needed} ) ) {
             $Kernel::OM->Get('Kernel::System::Log')->Log(
@@ -885,17 +888,55 @@ sub _PDFOutputCustomerInfos {
         }
     }
 
-    my $LayoutObject = $Kernel::OM->Get('Kernel::Output::HTML::Layout');
-
+    FIELD:
     for my $Field ( @{$Map} ) {
-        if ( ${$Field}[3] && $CustomerData{ ${$Field}[0] } ) {
-            $TableParam{CellData}[$Row][0]{Content} = $LayoutObject->{LanguageObject}->Translate( ${$Field}[1] ) . ':';
-            $TableParam{CellData}[$Row][0]{Font}    = 'ProportionalBold';
-            $TableParam{CellData}[$Row][1]{Content} = $CustomerData{ ${$Field}[0] };
+        next FIELD if !defined $CustomerData{ $Field->[0] };
+        next FIELD if !length $CustomerData{ $Field->[0] };
+        next FIELD if !$Field->[3];                            # not visible
 
-            $Row++;
-            $Output = 1;
+        my $Label = $LayoutObject->{LanguageObject}->Translate( $Field->[1] ) . ':';
+        my $Value = $CustomerData{ $Field->[0] };
+
+        # render dynamic field values
+        if ( $Field->[5] eq 'dynamic_field' ) {
+            next FIELD if $Field->[0] !~ m{\ADynamicField_(.*)\z};
+
+            my $DynamicFieldName   = $1;
+            my $DynamicFieldConfig = $DynamicFieldObject->DynamicFieldGet(
+                Name => $DynamicFieldName,
+            );
+            next FIELD if !IsHashRefWithData($DynamicFieldConfig);
+
+            my $Values = $Value;
+            if ( !IsArrayRefWithData($Values) ) {
+                $Values = [$Value];
+            }
+
+            my @RenderedValues;
+            VALUE:
+            for my $Value ( @{$Values} ) {
+                my $RenderedValue = $DynamicFieldBackendObject->DisplayValueRender(
+                    DynamicFieldConfig => $DynamicFieldConfig,
+                    Value              => $Value,
+                    HTMLOutput         => 0,
+                    LayoutObject       => $LayoutObject,
+                );
+
+                next VALUE if !IsHashRefWithData($RenderedValue) || !defined $RenderedValue->{Value};
+
+                push @RenderedValues, $RenderedValue->{Value};
+            }
+
+            $Value = join ', ', @RenderedValues;
+            $Label = $DynamicFieldConfig->{Label};
         }
+
+        $TableParam{CellData}[$Row][0]{Font}    = 'ProportionalBold';
+        $TableParam{CellData}[$Row][0]{Content} = $Label;
+        $TableParam{CellData}[$Row][1]{Content} = $Value;
+
+        $Row++;
+        $Output = 1;
     }
     $TableParam{ColumnData}[0]{Width} = 80;
     $TableParam{ColumnData}[1]{Width} = 431;